requirements.txt
# === Core PyTorch Ecosystem ===
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# === Data Processing & Utilities ===
numpy>=1.21.0
tqdm>=4.65.0
Pillow>=9.0.0
matplotlib>=3.5.0

# === Metrics & Losses ===
lpips>=0.1.4